some needed castings
git-svn-id: file:///home/roberto/Git/svn@8266 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: c0839c01cf272d75668995dbcb19134c2eb24648

+++ trunk/xmipp/applications/programs/classify_kerdensom/classify_kerdensom_main.cpp
@@ -202,7 +202,7 @@ public:
         	id=MDsummary.addObject();
         	MDsummary.setValue(MDL_REF,(int)i,id);
         	MDsummary.setValue(MDL_CLASSIFICATION_INTRACLASS_DISTANCE,myMap->aveDistances[i],id);
-        	MDsummary.setValue(MDL_COUNT,myMap->classifSizeAt(i),id);
+        	MDsummary.setValue(MDL_COUNT,(size_t)myMap->classifSizeAt(i),id);
         }
         MDsummary.write(formatString("KerDenSOM_Cluster_Summary@%s",fn_out.c_str()),MD_APPEND);
 
@@ -244,7 +244,7 @@ public:
         vectorHeaderOut.setValue(MDL_YSIZE,size,idOut);
         vectorHeaderIn.getValue(MDL_ZSIZE,size,idIn);
         vectorHeaderOut.setValue(MDL_ZSIZE,size,idOut);
-        vectorHeaderOut.setValue(MDL_COUNT,myMap->size(),idOut);
+        vectorHeaderOut.setValue(MDL_COUNT,(size_t)myMap->size(),idOut);
         vectorHeaderIn.getValue(MDL_CLASSIFICATION_DATA_SIZE,vectorSize,idIn);
         vectorHeaderOut.setValue(MDL_CLASSIFICATION_DATA_SIZE,vectorSize,idOut);
         vectorHeaderOut.write(formatString("vectorHeader@%s",fn_out.c_str()),MD_APPEND);
